+++ src/component/SingleProduct.jsx
@@ -1,14 +1,16 @@
-import { useEffect, useState } from "react";
 import Rating from "./Rating"
 import { NavLink } from "react-router-dom"
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
 import { toast } from 'react-toastify';
 import { LOCALSTORAGE_CART_KEY, LOCALSTORAGE_AUTH_USER_KEY } from "../constant";
+import { CartContext } from "../contexts/CartContext";
+import { useContext } from "react";
 
 
 export default function SingleProduct({ productInfo }) {
-  // const [cartProduct, setCartProduct] = useState([])
+  const { setCartItemCount } = useContext(CartContext);
+
   let navigate = useNavigate()
 
   let selectedProductHandler = async (productId) => {
@@ -29,13 +31,22 @@ export default function SingleProduct({ productInfo }) {
         let res = await axios.post(`${import.meta.env.VITE_API_BASE_URL}/carts/add`, cartRequestPayload)
         if (res.data) {
           let newProduct = res.data.products[0]
-          let cartItem = {}
+          let cartItem = []
           if (localStorage.getItem(LOCALSTORAGE_CART_KEY)) {
             cartItem = JSON.parse(localStorage.getItem(LOCALSTORAGE_CART_KEY))
           }
-          cartItem[newProduct.id] = newProduct
-          localStorage.setItem(LOCALSTORAGE_CART_KEY, JSON.stringify(cartItem))
-          
+
+          let exists = false
+          for (let index = 0; index < cartItem.length; index++) {
+            if (newProduct.id == cartItem[index].id) {
+              exists = true
+            }
+          }
+          if (!exists) {
+            cartItem.push(newProduct)
+            localStorage.setItem(LOCALSTORAGE_CART_KEY, JSON.stringify(cartItem))
+            setCartItemCount(cartItem.length)
+          }
           toast.success('1 new item(s) have been added to your cart!')
           console.log(cartItem);
         }
